Add SkipList with Lockfree Reads
This implementation of SkipList is inspired by LevelDB implementation.
Writes have to be synchronized between threads, but reads are lock free.
Writes are appends only i.e., there is no delete or update supported.
The memory for every node in the list is kept until SkipList is
not destroyed.
Caveats:
  1. Same key cannot be inserted twice.
  2. No update or deletes supported.
  3. Writes need to be synchronized, but reads are lock free.

Tests:
1. UTs are added for the single threaded case
2. Simple UT added for threads.

TODO:
1. Free up the memory
2. More tests for concurrent skipLists.